✅ refactor out test gen
