+++ bananalyzer/runner/runner.py
@@ -1,13 +1,12 @@
 import os
 import tempfile
 from pathlib import Path
-from typing import Awaitable, Callable, Dict, List
-from urllib.parse import urlparse
+from typing import Awaitable, Callable, List
 
 import pytest
 from pydantic import BaseModel
 
-from bananalyzer.data.schemas import Eval, Example
+from bananalyzer.data.schemas import Example
 from bananalyzer.schema import AgentRunnerClass, PytestArgs
 
 TestType = Callable[[], Awaitable[None]]
@@ -16,58 +15,6 @@ TestType = Callable[[], Awaitable[None]]
 class BananalyzerTest(BaseModel):
     code: str
     example: Example
-
-
-class TestGenerator:
-    def __init__(self) -> None:
-        self._classnames: Dict[str, int] = {}
-
-    def generate_test(self, example: Example) -> BananalyzerTest:
-        return BananalyzerTest(
-            code=f"""
-@pytest.mark.asyncio
-class {self._generate_class_name(example)}:
-
-    @classmethod
-    def setup_class(cls):
-        cls.example = get_example_by_url("{example.url}")
-
-
-    @pytest_asyncio.fixture(scope="class")
-    async def result(self, context, agent):
-        yield await agent.run(context, self.example)
-
-    {"".join(self._generate_eval_test(eval_, i) for i, eval_ in enumerate(example.evals))}
-""",
-            example=example,
-        )
-
-    def _generate_eval_test(self, eval_: Eval, i: int) -> str:
-        if eval_.type == "json_match" and isinstance(eval_.expected, dict):
-            return f"""
-    @pytest.mark.parametrize("key", {list(eval_.expected.keys())})
-    async def test_match_field(self, key, result) -> None:
-        assert self.example.evals[{i}].expected.get(key, None) == result.get(key, None)
-
-"""
-        return f"""
-    async def test_{eval_.type}(self, result) -> None:
-        self.example.evals[{i}].eval_results(None, result)
-
-"""
-
-    def _generate_class_name(self, example: Example) -> str:
-        domain = urlparse(example.url).netloc
-        domain = domain.replace(".", "_")
-        if domain.startswith("www_"):
-            domain = domain[4:]
-
-        domain = "".join([part.capitalize() for part in domain.split("_")])
-
-        key = f"{example.type.capitalize()}{domain}"
-        self._classnames[key] = self._classnames.get(key, -1) + 1
-        suffix = "" if not self._classnames[key] else f"{self._classnames[key]}"
-        return f"Test{key}{suffix}"
 
 
 def create_test_file(
