+++ tests/test_generator.py
@@ -1,0 +1,105 @@
+from bananalyzer import Example
+from bananalyzer.data.schemas import Eval
+from bananalyzer.runner.generator import TestGenerator
+
+
+def test_generate_test_with_single_eval() -> None:
+    generator = TestGenerator()
+    example = Example(
+        id="1",
+        url="https://www.test.com",
+        source="mhtml",
+        domain="test",
+        subdomain="www",
+        type="fetch",
+        evals=[Eval(type="json_match", expected={"key": "value"})],
+        fetch_id="job_posting",
+        goal=None,
+    )
+    test = generator.generate_test(example)
+    assert "test_match_field" in test.code
+
+
+def test_generate_test_with_multiple_evals() -> None:
+    generator = TestGenerator()
+    example = Example(
+        id="1",
+        url="https://www.test.com",
+        source="mhtml",
+        domain="test",
+        subdomain="www",
+        type="fetch",
+        evals=[
+            Eval(type="json_match", expected={"key": "value"}),
+            Eval(type="end_url_match", expected="https://www.test.com"),
+        ],
+        fetch_id="job_posting",
+        goal=None,
+    )
+    test = generator.generate_test(example)
+    assert "test_match_field" in test.code
+    assert "test_end_url_match" in test.code
+
+
+def test_generate_class_name_without_www() -> None:
+    generator = TestGenerator()
+    example = Example(
+        id="1",
+        url="https://test.com",
+        source="mhtml",
+        domain="test",
+        subdomain="www",
+        type="fetch",
+        evals=[],
+        fetch_id="job_posting",
+        goal=None,
+    )
+    class_name = generator._generate_class_name(example)
+    assert class_name == "TestFetchTestCom"
+
+
+def test_generate_class_name_with_www() -> None:
+    generator = TestGenerator()
+    example = Example(
+        id="1",
+        url="https://www.test.com",
+        source="mhtml",
+        domain="test",
+        subdomain="www",
+        type="fetch",
+        evals=[],
+        fetch_id="job_posting",
+        goal=None,
+    )
+    class_name = generator._generate_class_name(example)
+    assert class_name == "TestFetchTestCom"
+
+
+def test_generate_class_name_with_multiple_domains() -> None:
+    generator = TestGenerator()
+    example1 = Example(
+        id="1",
+        url="https://test.com",
+        source="mhtml",
+        domain="test",
+        subdomain="www",
+        type="fetch",
+        evals=[],
+        fetch_id="job_posting",
+        goal=None,
+    )
+    example2 = Example(
+        id="2",
+        url="https://test.com",
+        source="mhtml",
+        domain="test",
+        subdomain="www",
+        type="fetch",
+        evals=[],
+        fetch_id="job_posting",
+        goal=None,
+    )
+    class_name1 = generator._generate_class_name(example1)
+    class_name2 = generator._generate_class_name(example2)
+    assert class_name1 == "TestFetchTestCom"
+    assert class_name2 == "TestFetchTestCom2"
